Refactored the sidebar into its own directory
