+++ frontend/src/components/Layout/Sidebar/SidebarContent.jsx
@@ -10,9 +10,9 @@ import FormLabel from "@material-ui/core/FormLabel";
 import FormControl from "@material-ui/core/FormControl";
 import FormGroup from "@material-ui/core/FormGroup";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
-import { MAX_DATE, MIN_DATE } from "./Datetime/Range";
-import Presets from "./Datetime/Presets"
-import Days from "./Days";
+import { MAX_DATE, MIN_DATE } from "../Datetime/Range";
+import Presets from "../Datetime/Presets"
+import Days from "../Days";
 
 
 class SidebarContent extends React.Component {
